requirements.txt
# To create the dataset (create_dataset.py)
python-dotenv
pandas
xlrd
pyarrow